# docs/contracts/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactType": {
        "enum": [
          "REFERENCE_IMAGE",
          "IMAGE",
          "AUDIO",
          "VIDEO",
          "STORYBOARD",
          "MANIFEST",
          "REVIEW_PACKAGE",
          "FINAL_PACKAGE",
          "PROVIDER_DIAGNOSTIC"
        ],
        "title": "ArtifactType",
        "type": "string"
      },
      "CampaignArtifactsResponse": {
        "additionalProperties": false,
        "properties": {
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "campaign_version": {
            "minimum": 1.0,
            "title": "Campaign Version",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicArtifactReference"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "campaign_id",
          "campaign_version",
          "items"
        ],
        "title": "CampaignArtifactsResponse",
        "type": "object"
      },
...
      "CampaignCreationAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "campaign_id": {
            "format": "uuid",
            "title": "Campaign Id",
            "type": "string"
          },
          "campaign_version": {
            "minimum": 1.0,
            "title": "Campaign Version",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "links": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Links",
            "type": "object"
          },
          "progress_percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress Percent",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/CampaignStatus"
          }
        },
        "required": [
          "campaign_id",
          "campaign_version",
          "job_id",
          "status",
          "progress_percent",
          "links"
        ],
        "title": "CampaignCreationAcceptedResponse",
        "type": "object"
      },
      "CampaignCreationRequest": {
        "additionalProperties": false,
        "properties": {
          "brand_colors": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "title": "Brand Colors",
            "type": "array"
          },
          "business_description": {
            "maxLength": 2000,
            "minLength": 20,
            "title": "Business Description",
            "type": "string"
          },
          "business_name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Business Name",
            "type": "string"
          },
          "call_to_action": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call To Action"
          },
          "campaign_goal": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Campaign Goal",
            "type": "string"
          },
          "key_message": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key Message"
          },
          "language": {
            "maxLength": 20,
            "minLength": 2,
            "title": "Language",
            "type": "string"
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Platforms",
            "type": "array"
          },
          "product_or_service": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Product Or Service",
            "type": "string"
          },
          "reference_artifact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Artifact Id"
          },
          "target_audience": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "tone": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Tone",
            "type": "string"
          }
        },
        "required": [
          "business_name",
          "product_or_service",
          "business_description",
          "campaign_goal",
          "platforms",
          "tone",
          "language"
        ],
        "title": "CampaignCreationRequest",
        "type": "object"
      },
...
      "CampaignListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CampaignSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "CampaignListResponse",
        "type": "object"
      },
...
      "CancellationRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "CancellationRequest",
        "type": "object"
      },
...
      "ConflictError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SanitizedWorkflowError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ConflictError",
        "type": "object"
      },
...
      "NotFoundError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SanitizedWorkflowError"
          }
        },
        "required": [
          "error"
        ],
        "title": "NotFoundError",
        "type": "object"
      },
...
      "StandardValidationError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/SanitizedWorkflowError"
          }
        },
        "required": [
          "error"
        ],
        "title": "StandardValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Campaign Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/campaigns": {
      "get": {
        "operationId": "list_campaigns_api_v1_campaigns_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Campaigns",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "create_campaign_api_v1_campaigns_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignCreationAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "operationId": "get_campaign_api_v1_campaigns__campaign_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/artifacts": {
      "get": {
        "operationId": "get_campaign_artifacts_api_v1_campaigns__campaign_id__artifacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ArtifactType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignArtifactsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Campaign Artifacts",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/events": {
      "get": {
        "operationId": "get_campaign_events_api_v1_campaigns__campaign_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Campaign Events",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/versions/{version}/approve": {
      "post": {
        "operationId": "approve_campaign_api_v1_campaigns__campaign_id__versions__version__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Approve Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/versions/{version}/cancel": {
      "post": {
        "operationId": "cancel_campaign_api_v1_campaigns__campaign_id__versions__version__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancellationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancellationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancellationResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/versions/{version}/retry": {
      "post": {
        "operationId": "retry_campaign_api_v1_campaigns__campaign_id__versions__version__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Retry Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/versions/{version}/revisions": {
      "post": {
        "operationId": "revise_campaign_api_v1_campaigns__campaign_id__versions__version__revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardValidationError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Revise Campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  }
}